rename and remove paged request
